feat(credential-providers): collect credential providers in single package (#2672)
* feat(credential-providers): collect credential providers in public facing package

* feat(credential-providers): support temporary credentials provider

Equivalent to ChainableTemporaryCredentials class in v2 SDK. The provider
is not re-exported from standalone package or used in the credential-
provider-ini because the roleAssumers function interface in other
packages does not suite the usecase of this provider. In roleAssumer
interface, the master credential is required. It's not true for
fromTemporaryCredentials, where master credentials can be skipped,
in which case the default credential will be used.

* docs: remove inline import to improve docs

* docs(credentail-providers): wrap other providers and add README

* docs(credential-provider): mark other provider as internal

* docs(upgrading): update the upgrading guide

* docs: use clientConfig to override all creds providers client config

Co-authored-by: Trivikram Kamat <16024985+trivikr@users.noreply.github.com>